create case function added

+++ src/main/java/com/fedag/CSR/controller/WinChanceController.java
@@ -4,6 +4,8 @@ import com.fedag.CSR.service.impl.WinChanceServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/api/v1")
 @RequiredArgsConstructor
@@ -12,7 +14,7 @@ public class WinChanceController{
     private final WinChanceServiceImpl winChanceService;
 
     @GetMapping("/spin/{id}")
-    public Long spinPack(@PathVariable Long id, @RequestHeader("Authorization") String userToken) {
+    public BigDecimal spinPack(@PathVariable BigDecimal id, @RequestHeader("Authorization") String userToken) {
         return winChanceService.spinCase(id, userToken);
     }
