+++ src/main/java/com/fedag/CSR/service/WinChanceService.java
@@ -1,5 +1,8 @@
 package com.fedag.CSR.service;
 
+import java.math.BigDecimal;
+
 public interface WinChanceService {
-    Long spinCase(Long id, String userToken);
+    BigDecimal spinCase(BigDecimal id, String userToken);
+
 }
